# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EmailSecurity API
  version: "1.1"
  description: OpenAPI spec for EmailSecurity backend. Import this into Postman.
servers:
  - url: http://127.0.0.1:8000
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Login and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Exchange refresh for new access
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or revoked refresh token
...
  /auth/logout:
    post:
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        user:
          type: object
          properties:
            username:
              type: string
            roles:
              type: array
              items:
                type: string
        roles:
          type: array
          items:
            type: string
        expires_in:
          type: integer
          example: 3600
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 3600
...